refactor: 💡 organized helper functions on their own files
