+++ apps/web/components/molecules/user/conversationBody/message-list.tsx
@@ -1,29 +1,17 @@
+/**
+ * Renders a list of messages in the conversation body.
+ * @param messages - An array of Message objects to be displayed.
+ * @param userImage - The URL of the image for the user's profile.
+ * @param providerImage - The URL of the image for the provider's profile.
+ * @returns A JSX element containing the list of messages.
+ */
+
 "use client";
 import React from 'react';
 import { Message } from 'ai';
 import MessageItem from '@/components/molecules/user/conversationBody/message-item';
 import { Divider } from '@nextui-org/react';
 
-/**
- * Props for the MessageList component
- * @interface
- */
-/*
-interface MessageListProps {
-    // Array of messages to be displayed 
-    messages: Message[];
-    // Image URL for the sender of the message 
-    senderImage: string;
-}
-*/
-
-/**
- * Component that displays a list of messages
- * @param {MessageListProps} props - Component props
- * @param {Message[]} props.messages - Array of messages to be displayed
- * @param {string} props.senderImage - Image URL for the sender of the message
- * @returns {JSX.Element} - Rendered component
- */
 export default function MessageList({ messages, userImage, providerImage }: { messages: Message[], userImage: string, providerImage: string }): JSX.Element {
     return (
         <>
